--- src/main/java/com/jupjup/www/jupjup/common/response/ErrorResponse.java
@@ -1,18 +1,0 @@
-package com.jupjup.www.jupjup.common.response;
-
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.*;
-import org.springframework.http.HttpStatus;
-
-@Setter @Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class ErrorResponse {
-    @JsonProperty
-    private String errorMessage;
-    @JsonProperty
-    private HttpStatus errorCode;
-
-}
